whatsapp-agent/requirements.txt
anthropic>=0.18.0
httpx>=0.27.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
